Update booking service application with new features and improvements
This commit includes a series of changes to the booking service application. The `BookingServices.Application.csproj` file was modified to remove references to certain folders. Several files were updated to include new properties and methods, such as `RestaurantId` and `Note` properties in `Bookings.cs`, `AddBookingRequest.cs`, and `BookingDTO.cs`. The `BookingServices.cs` file was updated with changes to the `BookingServices` class and modifications to the `GetAllBookingAsync`, `GetBookingByIdAsync`, and `UpdateBookingAsync` methods. New API endpoints were added in `BookingController.cs` and `BookingMenuController.cs`. New files were added to define commands for creating or updating booking menus and handling the removal and updating of bookings. The database model was updated with new indexes, properties, and password changes. A new migration was added to modify the `BookingMenu` and `Bookings` tables in the database, fix the foreign key in the `BookingMenu` table, add a `Note` column to the `Bookings` table, and update the password values for some users in the `Users` table. The migration also includes the addition of a new column `RestaurantId` into the `Bookings` table, creating a relationship with the `Restaurants` table.