Updated RestaurantMenuServices and RestaurantMenuDTO
Updated `GetRestaurantMenuByIdAsync` method in `RestaurantMenuServices.cs` to include related `MenuCategory` entity when fetching a `RestaurantMenu` entity, using Entity Framework's eager loading feature. This ensures that the `MenuCategory` related to a particular `RestaurantMenu` is also loaded from the database.

In `RestaurantMenuDTO.cs`, the `BookingMenu` property has been commented out from the `RestaurantMenuDTO` class, meaning it will no longer be a part of the `RestaurantMenuDTO` objects. This could be a temporary change for testing purposes or because the `BookingMenu` is not needed in the data transfer object.