Updated booking methods, nullable properties, and method parameters
Updated `BookingRemoveAllCommandHandler.cs` to change the method of getting all bookings and the method of removing all bookings. Made the `Picture` property nullable in `LoginUserByGoogleQuery.cs`. Updated the `UpdateMenuCategory` method in `MenuCategoryController.cs` to take the `id` parameter from the method parameters instead of from the query string. Added the `Picture` property to the `LoginUserByGoogleQuery` request in `UserController.cs`. Updated the `History` property in `FolderProfile.pubxml.user` with new timestamps.