Update booking service application with new features and improvements
This commit includes a series of updates and improvements to the booking service application. The changes include the addition of the `EmailConfirmed` property to the `Customers` class and the `AddCustomerRequest` object, and the modification of parsing `bookingId` in `IPNPaymentCommandHandler.cs`. The `BookingServices.Core` namespace was added to `CustomerServices.cs` and `UserServices.cs`, and the methods `GetAllCustomersAsync` and `GetAllUser` were updated. A new `HashPassword` method was added to `Utils.cs`, and several properties in `Transaction.cs` were changed to nullable. The `vnp_TxnRef` request data in `VnPayServices.cs` was modified, and new properties were added to several classes. The `CreatePayment` and `IPN` methods in `PaymentController.cs` were updated to be asynchronous, and new methods were added to `CustomerController.cs` and `UserController.cs`. New classes were added to handle email confirmation and user registration. A new migration was added to include the `EmailConfirmed` property in the `Customers` table. The changes also include a migration in Entity Framework Core, adding a new field `EmailConfirmed` to the `Customers` entity, creating several entities, and configuring the model that resulted from the migration. The `Transaction` and `Users` tables in the `BookingServices` database were updated.